+++ aptos-move/aptos-transaction-benchmarks/src/main.rs
@@ -84,6 +84,9 @@ struct ExecuteOpt {
 
     #[clap(long)]
     pub maybe_block_gas_limit: Option<u64>,
+
+    #[clap(long, default_value_t = false)]
+    pub generate_then_execute: bool,
 }
 
 fn param_sweep(opt: ParamSweepOpt) {
@@ -121,6 +124,7 @@ fn param_sweep(opt: ParamSweepOpt) {
                 None,
                 false,
                 maybe_block_gas_limit,
+                false,
             );
             par_tps.sort();
             seq_tps.sort();
@@ -169,6 +173,7 @@ fn param_sweep(opt: ParamSweepOpt) {
 }
 
 fn execute(opt: ExecuteOpt) {
+    disable_speculative_logging();
     let bencher = TransactionBencher::new(any_with::<P2PTransferGen>((1_000, 1_000_000)));
 
     let (par_tps, _) = bencher.blockstm_benchmark(
@@ -181,8 +186,10 @@ fn execute(opt: ExecuteOpt) {
         opt.num_executor_shards,
         opt.concurrency_level_per_shard,
         opt.remote_executor_addresses,
-        opt.no_conflict_txns,
+        false,
+        //opt.no_conflict_txns,
         opt.maybe_block_gas_limit,
+        opt.generate_then_execute,
     );
 
     let sum: usize = par_tps.iter().sum();
